+++ tests/Hyperwallet/Tests/HyperwalletTest.php
@@ -319,6 +319,38 @@ class HyperwalletTest extends \PHPUnit_Framework_TestCase {
     }
 
     //--------------------------------------
+    // Client Token
+    //--------------------------------------
+
+    public function testGetClientToken_noUserToken() {
+        // Setup
+        $client = new Hyperwallet('test-username', 'test-password');
+
+        try {
+            $client->getClientToken('');
+            $this->fail('HyperwalletArgumentException expected');
+        } catch (HyperwalletArgumentException $e) {
+            $this->assertEquals('userToken is required!', $e->getMessage());
+        }
+    }
+
+    public function testGetClientToken_allParameters() {
+        // Setup data
+        $client = new Hyperwallet('test-username', 'test-password', 'test-program-token');
+        $apiClientMock = $this->createAndInjectApiClientMock($client);
+
+        \Phake::when($apiClientMock)->doPost('/rest/v3/users/{user-token}/client-token', array('user-token' => 'test-user-token'), null, array())->thenReturn(array('value' => 'true'));
+
+        // Run test
+        $clientToken = $client->getClientToken('test-user-token');
+        $this->assertNotNull($clientToken);
+        $this->assertEquals(array('value' => 'true'), $clientToken->getProperties());
+
+        // Validate mock
+        \Phake::verify($apiClientMock)->doPost('/rest/v3/users/{user-token}/client-token', array('user-token' => 'test-user-token'), null, array());
+    }
+
+    //--------------------------------------
     // Paper Checks
     //--------------------------------------